pyproject.toml
[project]
name = "langgraph-llm-tool-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pydantic>=2.5.0",
    "gradio>=4.19.0",
    "openai>=1.12.0",
    "python-dotenv>=1.1.0",
    "langchain-community>=0.0.20",
    "langgraph-checkpoint-sqlite>=2.0.10",
]
